+++ core/server/api_container/main.go
@@ -33,6 +33,9 @@ import (
 	"google.golang.org/grpc"
 	"net"
 	"os"
+	"path"
+	"runtime"
+	"strings"
 	"time"
 )
 
@@ -46,6 +49,10 @@ const (
 
 	forceColors   = true
 	fullTimestamp = true
+
+	logMethodAlongWithLogLine = true
+	functionPathSeparator     = "."
+	emptyFunctionName         = ""
 )
 
 type doNothingMetricsClientCallback struct{}
@@ -54,6 +61,8 @@ func (d doNothingMetricsClientCallback) Success()          {}
 func (d doNothingMetricsClientCallback) Failure(err error) {}
 
 func main() {
+	// This allows the filename & function to be reported
+	logrus.SetReportCaller(logMethodAlongWithLogLine)
 	// NOTE: we'll want to change the ForceColors to false if we ever want structured logging
 	logrus.SetFormatter(&logrus.TextFormatter{
 		ForceColors:               forceColors,
@@ -70,7 +79,12 @@ func main() {
 		PadLevelText:              false,
 		QuoteEmptyFields:          false,
 		FieldMap:                  nil,
-		CallerPrettyfier:          nil,
+		CallerPrettyfier: func(f *runtime.Frame) (string, string) {
+			fullFunctionPath := strings.Split(f.Function, functionPathSeparator)
+			functionName := fullFunctionPath[len(fullFunctionPath)-1]
+			_, filename := path.Split(f.File)
+			return emptyFunctionName, formatFilenameFunctionForLogs(filename, functionName)
+		},
 	})
 
 	err := runMain()
@@ -269,4 +283,14 @@ func createServiceNetworkAndModuleStore(
 	moduleStore := module_store.NewModuleStore(enclaveId, kurtosisBackend, moduleLauncher)
 
 	return serviceNetwork, moduleStore, nil
+}
+
+func formatFilenameFunctionForLogs(filename string, functionName string) string {
+	var output strings.Builder
+	output.WriteString("[")
+	output.WriteString(filename)
+	output.WriteString(":")
+	output.WriteString(functionName)
+	output.WriteString("]")
+	return output.String()
 }