+++ docs/changelog.md
@@ -3,6 +3,9 @@
 ### Changes
 - Updated `datastore-army-module` -> `datastore-army-package`
 
+### Features
+- Log file name and function like [filename.go:FunctionName()] while logging in `core` & `engine`
+
 # 0.57.1
 
 ### Changes